Began Prototyping first Level
I added the crate and room models into the game. I started setting up
the first room of the level. The turret script is currently being
programed it sense whether robit is to the left or right of it but the
rotation of the turret is a having some problems. I will aim to fix this
in next commit.